[Refactor] Moving RuntimeFilter into ExecCore
