+++ be/src/connector/connector.cpp
@@ -26,7 +26,7 @@
 #include "connector/jdbc_connector.h"
 #include "connector/lake_connector.h"
 #include "connector/mysql_connector.h"
-#include "runtime/runtime_filter/runtime_filter_helper.h"
+#include "exec/runtime_filter/runtime_filter_helper.h"
 
 namespace starrocks::connector {
