+++ be/src/exec/cross_join_node.h
@@ -17,10 +17,10 @@
 #include "column/vectorized_fwd.h"
 #include "common/statusor.h"
 #include "exec/pipeline_node.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
 #include "exprs/expr_context.h"
 #include "gen_cpp/PlanNodes_types.h"
 #include "runtime/descriptors.h"
-#include "runtime/runtime_filter/runtime_filter_descriptor.h"
 #include "runtime/runtime_state_fwd.h"
 
 namespace starrocks {
