+++ be/src/exec/pipeline/hashjoin/hash_join_build_operator.cpp
@@ -22,9 +22,9 @@
 #include "exec/pipeline/hashjoin/hash_joiner_factory.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/runtime_filter_types.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "runtime/current_thread.h"
-#include "runtime/runtime_filter/runtime_filter_descriptor.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 #include "runtime/runtime_filter_worker.h"
 
 namespace starrocks::pipeline {
