+++ be/src/exec/pipeline/operator.h
@@ -21,11 +21,11 @@
 #include "common/statusor.h"
 #include "exec/pipeline/runtime_filter_types.h"
 #include "exec/pipeline/schedule/observer.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "exec/spill/operator_mem_resource_manager.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/descriptors.h"
 #include "runtime/mem_tracker.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 
 namespace starrocks {
 class Expr;
