+++ be/src/exec/pipeline/pipeline_driver.h
@@ -34,10 +34,10 @@
 #include "exec/pipeline/schedule/observer.h"
 #include "exec/pipeline/schedule/pipeline_timer.h"
 #include "exec/pipeline/source_operator.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "fmt/printf.h"
 #include "runtime/mem_tracker.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 
 namespace starrocks {
