+++ be/src/exec/pipeline/scan/olap_scan_context.cpp
@@ -20,8 +20,8 @@
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/scan/glm_manager.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "runtime/global_dict/fragment_dict_state.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 #include "runtime/runtime_state_helper.h"
 #ifdef STARROCKS_JIT_ENABLE
 #include "exprs/jit/expr_jit_pass.h"
