+++ be/src/exec/pipeline/sort/sort_context.cpp
@@ -19,13 +19,13 @@
 #include <utility>
 
 #include "column/vectorized_fwd.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "exec/sorting/merge.h"
 #include "exec/sorting/sorting.h"
 #include "runtime/chunk_cursor.h"
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
-#include "runtime/runtime_filter/runtime_filter_descriptor.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 
 namespace starrocks::pipeline {
