+++ be/src/exec/pipeline/sort/sort_context.h
@@ -26,10 +26,10 @@
 #include "exec/pipeline/context_with_dependency.h"
 #include "exec/pipeline/runtime_filter_types.h"
 #include "exec/pipeline/schedule/observer.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "exec/sorting/merge.h"
 #include "exec/sorting/sorting.h"
-#include "runtime/runtime_filter/runtime_filter_descriptor.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 
 namespace starrocks::pipeline {
