+++ be/src/exec/runtime_filter/runtime_filter_descriptor.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runtime/runtime_filter/runtime_filter_descriptor.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
 
 #include "exprs/expr_factory.h"
