+++ be/src/formats/orc/utils.h
@@ -20,10 +20,10 @@
 #include "cctz/civil_time.h"
 #include "cctz/time_zone.h"
 #include "column/vectorized_fwd.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "formats/orc/orc_mapping.h"
 #include "gen_cpp/orc_proto.pb.h"
 #include "io/shared_buffered_input_stream.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 #include "types/date_value.h"
 #include "types/timestamp_value.h"
