+++ be/src/runtime/runtime_filter_worker.cpp
@@ -27,13 +27,13 @@
 #include "common/system/backend_options.h"
 #include "common/thread/thread.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "gen_cpp/Types_types.h" // for TUniqueId
 #include "gen_cpp/internal_service.pb.h"
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
 #include "runtime/fragment_mgr.h"
-#include "runtime/runtime_filter/runtime_filter_descriptor.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 #include "runtime/runtime_filter_builder.h"
 #include "runtime/runtime_filter_cache.h"
 #include "runtime/runtime_filter_factory.h"
