+++ be/src/storage/CMakeLists.txt
@@ -312,3 +312,5 @@ endif()
 ADD_BE_LIB(Storage
     ${STORAGE_FILES}
 )
+
+target_link_libraries(Storage PUBLIC ExecCore)
