+++ be/src/storage/runtime_range_pruner.hpp
@@ -19,8 +19,8 @@
 #include <utility>
 
 #include "exec/olap_common.h"
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "runtime/global_dict/config.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 #include "runtime/runtime_in_filter.h"
 #include "storage/column_and_predicate.h"
 #include "storage/column_or_predicate.h"
