+++ be/test/CMakeLists.txt
@@ -365,6 +365,7 @@ add_subdirectory(base)
 add_subdirectory(common)
 add_subdirectory(column)
 add_subdirectory(types)
+add_subdirectory(exec)
 add_subdirectory(runtime)
 add_subdirectory(exprs)
 add_subdirectory(io)
