+++ be/test/formats/parquet/file_reader_test.cpp
@@ -34,6 +34,7 @@
 #include "common/logging.h"
 #include "common/util/thrift_util.h"
 #include "exec/hdfs_scanner/hdfs_scanner.h"
+#include "exec/runtime_filter/runtime_filter_helper.h"
 #include "exprs/binary_predicate.h"
 #include "exprs/expr_context.h"
 #include "exprs/expr_executor.h"
@@ -51,7 +52,6 @@
 #include "runtime/global_dict/fragment_dict_state.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_filter.h"
-#include "runtime/runtime_filter/runtime_filter_helper.h"
 #include "runtime/runtime_state.h"
 #include "testutil/column_test_helper.h"
 #include "testutil/exprs_test_helper.h"
