+++ be/test/storage/olap_runtime_range_pruner_test.cpp
@@ -14,10 +14,10 @@
 
 #include <gtest/gtest.h>
 
+#include "exec/runtime_filter/runtime_filter_probe.h"
 #include "gen_cpp/RuntimeFilter_types.h"
 #include "runtime/descriptors.h"
 #include "runtime/runtime_filter.h"
-#include "runtime/runtime_filter/runtime_filter_probe.h"
 #include "runtime/runtime_state.h"
 #include "storage/column_predicate.h"
 #include "storage/predicate_parser.h"
